requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
jinja2>=3.1.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
alembic>=1.12.0
passlib>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
fastapi-mail>=1.4.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
